Rename `operations` module since it no longer contains any operations.
